Adds timestamp to build preview
Adds timestamp from when a build was created to the build preview as a title attribute. This maintains the existing behavior of humanized, relative-to-now time, but on hover, reveals the full date and time in UTC.

I could not figure out yet how to make this time respect the user's timezone. I seek guidance on how to do that within Elm / Vela... I see that Main.elm appears to have this information in scope.

Resolves: https://github.com/go-vela/community/issues/58