Adds autofocus to home page search
Adds autofocus to the home page search input, so a user can start typing immediately upon authenticating to Vela and find the repo they are looking for among their favorites.

I tried adding a cypress test for this but two approaches both failed/timed out.

I recognize this is unsolicited UX work so no hard feelings will be felt if this doesn't align with your standards.