Adds zone from the main model down to the build to normalize the age against a user's timezone
With @davidvader's help within https://github.com/go-vela/ui/pull/223\#issuecomment-716629348 I was able to get this working finally. I added a single test to verify the title attr is present - but stopped short of attempting to assert on the value - I understand build.created is mocked but I am unsure about the timezone on a server and time is a scary thing to test against.