fix(logs): better ux with missing logs, prevent refetch
previously, if a user expanded a log for which the log call returned a
404, all the current steps would disappear and you would see a spinner
on the page under the tabbed navigation. that shouldn't happen if one
step's log was missing. this change will show a message in place of the
logs that explains that the logs likely have expired.

also included is a change that will prevent refetching logs for a step
that is in finite state and already has fetched logs.

lastly, there is a change that should fix
https://github.com/go-vela/community/issues/861